add integer value to integration test.

+++ pkg/sink/codec/simple/avro.go
@@ -40,14 +40,11 @@ func newTableSchemaMap(tableInfo *model.TableInfo) interface{} {
 			"mysqlType": types.TypeToStr(col.GetType(), col.GetCharset()),
 			"charset":   col.GetCharset(),
 			"collate":   col.GetCollate(),
-			// todo: add detail description about length,
-			// for the text type, default length is 4294967295,
-			// it's out of the range, convert it to int32, make it -1.
-			"length":   int32(col.GetFlen()),
-			"decimal":  col.GetDecimal(),
-			"elements": col.GetElems(),
-			"unsigned": mysql.HasUnsignedFlag(col.GetFlag()),
-			"zerofill": mysql.HasZerofillFlag(col.GetFlag()),
+			"length":    col.GetFlen(),
+			"decimal":   col.GetDecimal(),
+			"elements":  col.GetElems(),
+			"unsigned":  mysql.HasUnsignedFlag(col.GetFlag()),
+			"zerofill":  mysql.HasZerofillFlag(col.GetFlag()),
 		}
 		column := map[string]interface{}{
 			"name":     col.Name.O,
@@ -314,7 +311,7 @@ func newTableSchemaFromAvroNative(native map[string]interface{}) *TableSchema {
 			MySQLType: rawDataType["mysqlType"].(string),
 			Charset:   rawDataType["charset"].(string),
 			Collate:   rawDataType["collate"].(string),
-			Length:    int(rawDataType["length"].(int32)),
+			Length:    int(rawDataType["length"].(int64)),
 			Decimal:   int(rawDataType["decimal"].(int32)),
 			Elements:  elements,
 			Unsigned:  rawDataType["unsigned"].(bool),
